Changed abl delete recipe, added user authorization

+++ uu_kucharka_maing01-server/abl/recipe/delete-abl.js
@@ -2,6 +2,8 @@ const path = require("path");
 const Ajv = require("ajv").default;
 const dao = require("../../dao/recipe-dao");
 
+const userUtils = require("../UserUtils");
+
 let schema = {
   type: "object",
   properties: {
@@ -15,15 +17,27 @@ async function DeleteAbl(req, res) {
   const valid = ajv.validate(schema, req.body);
   try {
     if (valid) {
-      const recipeId = req.body.id;
-      await dao.deleteRecipe(recipeId);
-      res.json({});
-      if (!recipeId) {
+      const { recipeId, userId } = req.body;
+      const recipe = await dao.getRecipe(recipeId);
+
+      if (!recipe) {
         res
-          .status(400)
+          .status(404)
           .send({ error: `Recipe with id '${recipeId}' doesn't exist.` });
+        return;
       }
-      res.json(recipeId);
+
+      const isAdmin = userUtils.hasAuthority(userId, userUtils.ADMIN);
+      const isCreator = userUtils.hasAuthority(userId, userUtils.CREATOR);
+      const isAuthor = recipe.createdBy === userId;
+
+      if (isAdmin || (isCreator && isAuthor)) {
+        await dao.deleteRecipe(recipeId);
+
+        res.json({ recipe });
+      } else {
+        res.json({ message: "You are not allowed to delete this recipe" });
+      }
     } else {
       res.status(400).send({
         errorMessage: "validation of input failed",
